Creating components
Question component = Receive the questions functions.
Answers component = Receive the answers functions.

These components was created to easy manager the functionalities, and not generate a bugs to use the same 'key'. For this, the 'Question' component recive 'Answers' component on your structuring, because both the components use the same 'key' 'activeQuestionIndex' to manager yours functionalities.