Update a QuestionTimer.
In this commit was created a constant of name 'timer' in the Question component to receive a new values to control the 'timeout' method in QuestionTimer component.
Was added one 'key' in the call QuestionTimer component on Question component, to correct a bug when the timer value was changed.
Was also change the called 'onTimeout' in the Question component, so he calls the 'skipAnswer' if the answers was selected or not, kinda manager the call.